meeting/requirements.txt
Django==2.2.6
pytz==2019.3
whitenoise==4.1.4
gunicorn==20.0.0
psycopg2==2.8.4
djangorestframework==3.10.3
dj-config-url==0.1.1
django-heroku==0.3.1
django-crispy-forms==1.7.2
xhtml2pdf==0.2.4
